Reference

MPILargeCounts.MPILargeCounts — Module

Chunked communication based on MPI.jl with arbitrary-size data.

MPILargeCounts.Irecv! — Method

req = Irecv!(recvbuf, comm::Comm;
        source::Integer=MPI.ANY_SOURCE, tag::Integer=MPI.ANY_TAG)

Starts a nonblocking receive into the buffer recvbuf from MPI rank source of communicator comm using the message tag tag.

MPILargeCounts.Isend — Method

Isend(data, comm::Comm; dest::Integer, tag::Integer=0)

Starts a nonblocking send of data to MPI rank dest of communicator comm using with the message tag tag.

MPILargeCounts.Recv! — Function

data = Recv!(recvbuf, comm::Comm;
        source::Integer=MPI.ANY_SOURCE, tag::Integer=MPI.ANY_TAG)

Completes a blocking receive into the buffer recvbuf from MPI rank source of communicator comm using with the message tag tag.

MPILargeCounts.Send — Method

Send(buf, comm::Comm; dest::Integer, tag::Integer=0)

Perform a blocking send from the buffer buf to MPI rank dest of communicator comm using the message tag tag.

Send(obj, comm::Comm; dest::Integer, tag::Integer=0)

Complete a blocking send of an isbits object obj to MPI rank dest of communicator comm using with the message tag tag.

MPILargeCounts.allreduce — Method

allreduce(obj, op, comm::Comm; root::Integer=0)
allreduce!(obj, op, comm::Comm; root::Integer=0)

Performs elementwise reduction using the operator op on the object obj across all ranks in comm, returning the reduced object on all ranks.

MPILargeCounts.bcast — Method

bcast(obj, comm::Comm; root::Integer=0)

Broadcast the object obj from rank root to all processes on comm. This is able to handle arbitrary data.

MPILargeCounts.isend — Method

isend(obj, comm::Comm; dest::Integer, tag::Integer=0)

Starts a nonblocking send of using a serialized version of obj to MPI rank dest of communicator comm using with the message tag tag.

Returns the communication Request for the nonblocking send.

MPILargeCounts.mpi_execute_on_root — Method

mpi_execute_on_root(F::A, args...; kwargs...)

Execute the function F with arguments args... and keyword arguments kwargs... on the root of the communicator comm (default: MPI.COMM_WORLD). If blocking is set to true (default: false), a barrier is performed before and after the execution to synchronize all ranks. Returns the result of F on the root, and nothing on all other ranks.

MPILargeCounts.mpi_execute_on_root_and_bcast — Method

mpi_execute_on_root_and_bcast(F::A, args...; kwargs...)

Execute the function F with arguments args... and keyword arguments kwargs... on the root of the communicator comm (default: MPI.COMM_WORLD), and broadcast the result to all ranks. Broadcasts the result of F on all ranks.

MPILargeCounts.recv — Function

obj = recv(comm::Comm;
        source::Integer=MPI.ANY_SOURCE, tag::Integer=MPI.ANY_TAG)

Completes a blocking receive of a serialized object from MPI rank source of communicator comm using with the message tag tag.

MPILargeCounts.reduce — Method

reduce(obj, op, comm::Comm; root::Integer=0)
reduce!(obj, op, comm::Comm; root::Integer=0)

Performs elementwise reduction using the operator op on the object obj on the root rank root of comm, returning the reduced object on rank root.

MPILargeCounts.send — Method

send(obj, comm::Comm; dest::Integer, tag::Integer=0)

Complete a blocking send using a serialized version of obj to MPI rank dest of communicator comm using with the message tag tag.

MPILargeCounts.split_buffer — Method

split_buffer(vec)

Split an AbstractVector into a Vector{MPI.Buffer} views without copying.